optimize writing of the relative time
The relative time feature has strong requirements regarding data
matching (HW and host stamps must be related to the one time point).
But it is not necessary to obtain the latest available data. Based on
the above it is proposed to add an optimized data writing, when update
operation may be skipped if the data is locked.

To prevent starvation (when the writing operation is always skipped)
is added an obligatory waiting if maximal skip was achieved.

Note that this patch is controversial (seems like a premature
optimization), so it should be used carefully.

Signed-off-by: Gluttton <gluttton@ukr.net>