Fix configuration of Octave plugin with static lib
Fix configuration of the Octave plugin with static LimeSuite library
(with `BUILD_SHARED_LIBS=OFF` option). The current implementation uses
the `TARGET_SONAME_FILE_DIR` property that can be used only for shared
libraries. It is proposed to use the `TARGET_FILE_DIR` property that can
be used for shared and static libraries as well.

In fact, despite the configuration error, the Octave plugin can be
successfully built. So the current implementation doesn't affect the
build process when it is performed manually, but it may be a real
problem when LimeSute is integrated into CI process.

Test: to ensure that configuration is completed successfully:
      $ cmake -DBUILD_SHARED_LIBS=OFF ..
      $ echo $?
      to ensure that the Octave plugin isn't affected:
      $ cmake -DBUILD_SHARED_LIBS=OFF ..
      $ make LimeSuiteOctave VERBOSE=1
      and:
      $ cmake -DBUILD_SHARED_LIBS=ON ..
      $ make LimeSuiteOctave VERBOSE=1
Signed-off-by: Gluttton <gluttton@ukr.net>